new changes in migration

+++ database/migrations/2026_01_13_000545_add_template_to_pages_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('core_pages', function (Blueprint $table) {
-            $table->string('template')->nullable();
-        });
+        if (!Schema::hasColumn('core_pages', 'template')) {
+            Schema::table('core_pages', function (Blueprint $table) {
+                $table->string('template')->nullable();
+            });
+        }
     }
 
     /**
@@ -23,8 +25,10 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('core_pages', function (Blueprint $table) {
-            $table->dropColumn('template');
-        });
+        if (Schema::hasColumn('core_pages', 'template')) {
+            Schema::table('core_pages', function (Blueprint $table) {
+                $table->dropColumn('template');
+            });
+        }
     }
 };
